replaced cifar_small_sample.py with the contents of refactor_cifar_small_sample.py
